requirements.txt
elevenlabs==1.3.1
memgpt==0.1.0
openai==1.35.7
pygame==2.6.0
python-dotenv==1.0.1
